organize expr files into a directory
  - rename `pathexpr`, `boolexpr`, `exprparse` to `path_expr`,
    `bool_expr`, `expr_parser` respectively
  - handle requiring exprs centrally from within the `Impl` module
