+++ lib/wab/impl/bool_expr.rb
@@ -1,6 +1,4 @@
 
-require 'wab/impl/expr'
-
 module WAB
   module Impl
