+++ lib/wab/impl/expr.rb
@@ -1,6 +1,4 @@
 
-require 'wab'
-
 module WAB
   module Impl
