+++ lib/wab/impl/exprs/gte.rb
@@ -1,14 +1,12 @@
 
-require 'wab/impl/pathexpr'
-
 module WAB
   module Impl
 
-    # Matches a node that has a value greater than the provided value. If a
-    # integer or float is provided then both integer and floats are
-    # checked. If the value provided is a time then only time nodes are
+    # Matches a node that has a value greater than or equals to the provided
+    # value. If a integer or float is provided then both integer and floats
+    # are checked. If the value provided is a time then only time nodes are
     # checked. Any other type results in an error.
-    class Gt < PathExpr
+    class Gte < PathExpr
 
       # Creates a new instance with the provided parameters.
       #
@@ -20,13 +18,13 @@ module WAB
       end
 
       def eval(data)
-        data.get(@path) > @value
+        data.get(@path) >= @value
       end
 
       def native()
-        ['GT', @path, @value]
+        ['GTE', @path, @value]
       end
 
-    end # Gt
+    end # Gte
   end # Impl
 end # WAB
